+++ next.config.js
@@ -2,6 +2,7 @@ const withPWA = require("next-pwa");
 
 module.exports = withPWA({
   pwa: {
+    disable: process.env.NODE_ENV === 'development',
     dest: "public",
     register: true,
   },
