+++ src/components/mainSections/HeroSection.tsx
