+++ src/components/mainSections/index.tsx
@@ -1,0 +1,4 @@
+export { default as CountryData } from "./CountryData";
+export { default as GlobalData } from "./GlobalData";
+export { default as HeroSection } from "./HeroSection";
+export { default as IndonesiaData } from "./IndonesiaData";
