+++ src/components/stats/components/StatCard.tsx
@@ -1,5 +1,6 @@
 import { Box, Flex, Text } from "@chakra-ui/core";
-import formatNumber from "../../../helpers/formatNumber";
+
+import { formatNumber } from "../../../helpers";
 
 type StatCardProps = {
   value: number;
