+++ src/components/stats/components/index.tsx
@@ -1,0 +1,2 @@
+export { default as StatCard } from "./StatCard";
+export { default as StatWrap } from "./StatWrap";
