+++ src/components/stats/index.tsx
@@ -1,0 +1,3 @@
+export { default as CountryStat } from "./CountryStat";
+export { default as GlobalStat } from "./GlobalStat";
+export { default as LocalStat } from "./LocalStat";
