+++ src/helpers/dateFormat.ts
@@ -10,5 +10,4 @@ export const dateFormatLong = (date: string) => {
   };
 
   return convertedDate.toLocaleDateString("default", options);
-  // return `${convertedDate.getDay()}-${convertedDate.getMonth()}-${convertedDate.getFullYear()} ${convertedDate.getHours()}:${convertedDate.getMinutes()}`;
 };
